port(phase2): понтовость, XP and money are words too -- no conversions left
The last three of the eight widened fields, narrowed to what `src/save.rs`
already declared and what the record holds:

* `Game::pontovost_street` `i32` -> `i16` (`20ae:38cb`, a Pascal `Integer`;
  all 23 absolute references in the image are word-width, per the census in
  `docs/re/gaps.md`)
* `Progress::xp` / `threshold` `u32` -> `u16` (`20ae:38ce` / `20ae:38d0`;
  `1000:debe` is `add [0x38ce],ax`)
* `Fighter::money` `i32` -> `i16` (`20ae:38c7`)

**`Game::to_save` and `Game::from_save` now convert nothing.** All eight
fields assign straight across in both directions:

    grep -n 'as i16\|as i32\|i32::from' src/persist.rs | grep -cE \
      'money|junk|joint|beer|street_cred|armour|xp|threshold'
    # 0

Every write that the original does as a word op wraps here now --
`wrapping_add` / `wrapping_sub` at each store, rather than an implicit
widening that silently disagreed with the machine at the boundary. That
includes `progress::xp_award`, whose four-stat sum the original accumulates
in `ax` and adds as a word.

**What is NOT changed, deliberately.** Several gates compute in `i32` and
compare against the field (`(level_in_district - 5) * 5 + cred >= 0x28` and
its two siblings). The original computes those in `ax`, so they can wrap
mid-expression where this port cannot. Widening the field at those three
comparison sites keeps their behaviour exactly as it was; making the
EXPRESSIONS wrap is a separate question from the field's width and is not
smuggled in here.

Test fixtures read these words as words (`u16at(..) as i16`) rather than
widening them, and the deserialisation structs in `tests/progression.rs`,
`combat_sequence.rs` and `wander_sequence.rs` declare the record's widths.
No captured artifact was regenerated: `combat_trace.json` has always
recorded `"width": 2` for the purse fields and `"width": 1` for armour.

`cargo test` 460 passed, 0 failed. `difftest` 363/363. clippy and fmt clean.
`tools/test_string_citations.py` 10 passed.

+++ src/character_sheet.rs
@@ -125,9 +125,9 @@ pub const HEALTH_GREEN_ABOVE: f64 = 0.50;
 pub struct Kit {
     /// `20ae:38ce` -- XP not yet spent on a level
     /// (`crate::progress::Progress::xp`), pushed at `1000:1ab5`.
-    pub xp_38ce: u32,
+    pub xp_38ce: u16,
     /// `20ae:38d0` -- XP needed for the next level, pushed at `1000:1ab9`.
-    pub threshold_38d0: u32,
+    pub threshold_38d0: u16,
     /// `20ae:38cd` -- the joint buff's countdown. Read three times: the
     /// Сила colour slot (`1000:1acb`), the damage line's colour
     /// (`1000:1e06`) and the `Обдолбаный` condition (`1000:20ca`). All three
